Correct implementation by Jani to use ElapsedDays instead of Days, etc.
This class, which previously failed 6 out of 12 tests, was not the correct implementation given by Jani in his answer. In fact his answer as now given properly passes all tests (and was the first solution of the problem). We fixed the class to use properties ElapsedYears, ElapsedMonths, ElapsedDays instead of Years, Months, Days.